+++ NuGetPlus.Core/ProjectManagement.fs
@@ -45,6 +45,7 @@ let private getFiles project (package : IPackage) =
             "framework assemblies"
     let contentFiles = 
         grabCompatible project (package.GetContentFiles) "content files"
+        |> Seq.filter (fun p -> p.EffectivePath <> null && p.TargetFramework <> null)
     let buildFiles = 
         grabCompatible project (package.GetBuildFiles) "build files"
     assemblyReferences, frameworkReferences, contentFiles, buildFiles
@@ -93,9 +94,11 @@ let private RemoveFilesFromProj packageInstallPath (package : IPackage)
                                   (Path.GetExtension contentFile.EffectivePath))
                 |> Seq.distinctBy
                        (fun contentFile -> 
-                           contentFile.EffectivePath, 
-                           contentFile.TargetFramework.Version), 
-                
+                           let path =
+                                if String.IsNullOrEmpty contentFile.EffectivePath then "" else contentFile.EffectivePath
+                           let version =
+                                if contentFile.TargetFramework <> null then contentFile.TargetFramework.Version else Version()
+                           path, version), 
                 Seq.distinctBy 
                     (fun (buildFile : IPackageFile) -> 
                         buildFile.EffectivePath, 
@@ -206,6 +209,17 @@ let private UninstallFromPackagesConfigFile id (project : IProjectSystem) =
            configDoc.Element(XName.Get "packages") |> DeletePackageNode id
            SortPackages configDoc
            project.AddFile(fileName, fun (s : Stream) -> configDoc.Save(s))
+
+let HasReferenceToPackage id project =
+    let config = Path.Combine(Path.GetDirectoryName project, "packages.config")
+    if File.Exists config then
+        File.ReadAllText(config)
+        |> XDocument.Parse
+        |> (fun configDoc ->
+                configDoc.Element(XName.Get "packages").Elements()
+                |> Seq.exists (fun x -> x.Attribute(XName.Get "id") <> null && x.Attribute(XName.Get "id").Value = id))
+    else
+        false
 
 let private InstallToPackagesConfigFile (package : IPackage) 
     (project : IProjectSystem) =