+++ release_notes.md
@@ -1,3 +1,13 @@
+## Version 3.1.0
+
+New features:
+
+*	Added 'Scan' command to Solution Management
+
+Bug fixes:
+
+*	List on [the NuGetPlus YouTrack instance](http://nugetplus.myjetbrains.com/youtrack/issues?q=fixed+in%3A+3.1.0)
+
 ## Version 3.0.0
 
 Breaking changes: